fix: MathF.Clamp → float.Clamp in SmokeSegment; fix double→float cast in OsuModTargetPractice
Agent-Logs-Url: https://github.com/winnerspiros/osu/sessions/ce8f0232-7f26-43f2-bd5a-69bdb0e1b292

+++ osu.Game.Rulesets.Osu/Mods/OsuModTargetPractice.cs
@@ -450,7 +450,7 @@ namespace osu.Game.Rulesets.Osu.Mods
 
         private bool checkForOverlap(IEnumerable<OsuHitObject> objectsToCheck, OsuHitObject target)
         {
-            float radiusTwice = target.Radius * 2;
+            float radiusTwice = (float)target.Radius * 2;
             return objectsToCheck.Any(h => Vector2.DistanceSquared(h.Position, target.Position) < radiusTwice * radiusTwice);
         }
